+++ src/bos/operators/session_completion.py
@@ -22,6 +22,11 @@
 # ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR
 # OTHER DEALINGS IN THE SOFTWARE.
 #
+
+"""
+BOS session completion operator
+"""
+
 import logging
 
 from bos.common.clients.bos import BOSClient
@@ -39,10 +44,6 @@ class SessionCompletionOperator(BaseOperator):
     The Session Completion Operator marks sessions complete when all components
     that are part of the session have been disabled.
     """
-
-    @property
-    def name(self) -> str:
-        return 'SessionCompletion'
 
     # This operator overrides _run and does not use "filters" or "_act", but they are defined here
     # because they are abstract methods in the base class and must be implemented.