+++ src/bos/operators/utils/liveness/__main__.py
@@ -21,6 +21,11 @@
 # ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR
 # OTHER DEALINGS IN THE SOFTWARE.
 #
+
+"""
+Main entry point for liveness module
+"""
+
 import sys
 import logging
 import os